style: 릴리즈 노트 원고 수정 디자인 변경 및 오류 해결
