+++ src/components/ReleaseNote/ReleaseNoteDetail/SingleManuscript/styles.tsx
@@ -1,0 +1,5 @@
+import styled from '@emotion/styled';
+
+export const LastEditedMemberDiv = styled.div`
+  color: #00b050;
+`;
